+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.6 - 2021-05-10]
+
+### Added
+- Exponential Smoothing Tests.
+
+### Fixed
+- y-train on Forecaster run method.
+
 ## [0.3.5 - 2021-05-06]
 
 ### Changed